test-requirements.txt
-r requirements.txt
nose >= 1.3.0
django-nose >= 1.2
names >= 0.3.0